Restricted Key Space to UTF-8 Strings
If a byte array is encountered in the key part of a dictionary, we
it is converted to a string. There are better ways to handle edge cases
but the edge cases are insane (non UTF-8 keys why) and don't really work
well with Serde anyway.

This library should be burned to the ground and rewritten with nom. The
parsing is mixed with the serialization.